move Texture and Mesh to Model.h and Scene3D to Scene.h
