+++ lib/CMakeLists.txt
@@ -2,12 +2,9 @@
 set(HEADERS_LIST
   ${PROJECT_SOURCE_DIR}/include/RendGL/Camera.h
   ${PROJECT_SOURCE_DIR}/include/RendGL/Light.h
-  ${PROJECT_SOURCE_DIR}/include/RendGL/Mesh.h
   ${PROJECT_SOURCE_DIR}/include/RendGL/Model.h
   ${PROJECT_SOURCE_DIR}/include/RendGL/Scene.h
-  ${PROJECT_SOURCE_DIR}/include/RendGL/Scene3D.h
   ${PROJECT_SOURCE_DIR}/include/RendGL/Shader.h
-  ${PROJECT_SOURCE_DIR}/include/RendGL/Texture.h
   ${PROJECT_SOURCE_DIR}/include/RendGL/Utils.h
   ${PROJECT_SOURCE_DIR}/include/RendGL/Window.h
 )
@@ -15,12 +12,9 @@ set(HEADERS_LIST
 set(SOURCES_LIST
   ${PROJECT_SOURCE_DIR}/lib/Camera.cpp
   ${PROJECT_SOURCE_DIR}/lib/Light.cpp
-  ${PROJECT_SOURCE_DIR}/lib/Mesh.cpp
   ${PROJECT_SOURCE_DIR}/lib/Model.cpp
   ${PROJECT_SOURCE_DIR}/lib/Scene.cpp
-  ${PROJECT_SOURCE_DIR}/lib/Scene3D.cpp
   ${PROJECT_SOURCE_DIR}/lib/Shader.cpp
-  ${PROJECT_SOURCE_DIR}/lib/Texture.cpp
   ${PROJECT_SOURCE_DIR}/lib/Utils.cpp
   ${PROJECT_SOURCE_DIR}/lib/Window.cpp
 )
