--- lib/Mesh.cpp
@@ -1,86 +1,0 @@
-#include "Mesh.h"
-
-Mesh::Mesh(const std::vector<GLfloat>& vertices,
-    const std::vector<GLuint>& indices) 
-{
-    createMesh(vertices, indices);
-}
-
-Mesh::~Mesh()
-{
-    deleteMesh();
-}
-
-void Mesh::createMesh(const vector<GLfloat> & vertices,
-    const vector<GLuint> & indices)
-{
-    GLuint numVertices = vertices.size();
-    m_numIndices = indices.size();
-
-    // create a Vertex Array Object on the GPU and store its number
-    glGenVertexArrays(1, &m_VAO);
-    // activate VAO to add further objects to it
-    glBindVertexArray(m_VAO);
-
-    // create a Vertex Buffer Object on the GPU and store its number
-    glGenBuffers(1, &m_VBO);
-    // activate VBO
-    glBindBuffer(GL_ARRAY_BUFFER, m_VBO);
-    // copy vertices to the GPU
-    glBufferData(GL_ARRAY_BUFFER, sizeof(vertices[0]) * numVertices, &vertices[0], GL_STATIC_DRAW);
-
-    // we copy 5 values per vertex: the first 3 are the coordinates in 3D, the last 2 are the corresponding texture coordinates
-    // here we tell OpenGl which values are coordinates
-    // layout -> size of data (3 coordinates) -> type -> normalize -> stride -> offset
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(vertices[0]) * 8, 0);
-    // enable (layout = 0), no need to search to the vertices when using the vertex shader (?)
-    glEnableVertexAttribArray(0);
-
-    // here we tell OpenGL which values are texture coordinates
-    glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, sizeof(vertices[0]) * 8, (void*)(sizeof(vertices[0]) * 3));
-    // same as above
-    glEnableVertexAttribArray(1);
-    
-    // here we tell OpenGL which values are normal components
-    glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, sizeof(vertices[0]) * 8, (void*)(sizeof(vertices[0]) * 5));
-    // same as above
-    glEnableVertexAttribArray(2);
-
-    // create a Element Buffer Object on the GPU and store its number
-    glGenBuffers(1, &m_EBO);
-    // activate EBO
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_EBO);
-    // copy elements' indices to the GPU
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(indices[0]) * m_numIndices, &indices[0], GL_STATIC_DRAW);
-
-    // deactivate VAO, no need to edit this object (should it be here or after VBO and EBO?)
-    glBindVertexArray(0);
-    // deactivate VBO, no need to edit the vertex data any more
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
-    // deactivate EBO, no need to edit the index data any more
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER,0);
-}
-
-void Mesh::renderMesh() const
-{
-    // activate VAO, VAO = object (?)
-    glBindVertexArray(m_VAO);
-    // activate EBO
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER,m_EBO);
-    // 0 is a nullptr to smth
-    glDrawElements(GL_TRIANGLES, m_numIndices, GL_UNSIGNED_INT, nullptr);
-    // deactivate VAO and EBO
-    glBindVertexArray(0);
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER,0);
-}
-
-void Mesh::deleteMesh()
-{
-    // free GPU memory
-    if (m_EBO != 0)
-        glDeleteBuffers(1,&m_EBO);
-    if (m_VBO != 0)
-        glDeleteBuffers(1,&m_VBO);
-    if (m_VAO != 0)
-        glDeleteVertexArrays(1,&m_VAO);
-}
